Move calculate_toc_html to utils
