+++ lib/markdown2.py
@@ -112,7 +112,7 @@ import codecs
 from collections import defaultdict
 
 from lib.errors import MarkdownError
-from lib.utils import slugify
+from lib.utils import slugify, calculate_toc_html
 
 # ---- globals
 
@@ -2461,41 +2461,6 @@ class MarkdownWithExtras(Markdown):
 
 
 # ---- internal support functions
-
-
-def calculate_toc_html(toc):
-    """Return the HTML for the current TOC.
-
-    This expects the `_toc` attribute to have been set on this instance.
-    """
-    if toc is None:
-        return None
-
-    def indent():
-        return '  ' * (len(h_stack) - 1)
-    lines = []
-    h_stack = [0]   # stack of header-level numbers
-    for level, id, name in toc:
-        if level > h_stack[-1]:
-            lines.append("%s<ul>" % indent())
-            h_stack.append(level)
-        elif level == h_stack[-1]:
-            lines[-1] += "</li>"
-        else:
-            while level < h_stack[-1]:
-                h_stack.pop()
-                if not lines[-1].endswith("</li>"):
-                    lines[-1] += "</li>"
-                lines.append("%s</ul></li>" % indent())
-        lines.append('%s<li><a href="#%s">%s</a>' % (
-            indent(), id, name))
-    while len(h_stack) > 1:
-        h_stack.pop()
-        if not lines[-1].endswith("</li>"):
-            lines[-1] += "</li>"
-        lines.append("%s</ul>" % indent())
-    return '\n'.join(lines) + '\n'
-
 
 class UnicodeWithAttrs(str):
     """A subclass of unicode used for the return value of conversion to
